backend/requirements.txt

fastapi==0.124.2
uvicorn==0.38.0
pdfplumber==0.11.8
python-dotenv==1.2.1
langchain==1.2.0
langchain-text-splitters==1.1.0
scikit-learn==1.8.0
google-genai==1.56.0
python-multipart==0.0.9
pydantic-settings==2.2.1